challenge/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
SQLAlchemy==2.0.28
alembic==1.12.1
pydantic==2.5.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pytz
